Fix all repo source URLs
This will convert multiple types of URIs into HTTPS URLs that the browser
can understand. Repositories hosted by GitHub, GitLab, and Bitbucket are
all tested. If the repo URI isn't recognizable, just use the Forge page
for this module.

The end result is that everything displayed in a report should be
clickable in the browser.